+++ mollie.php
@@ -1279,7 +1279,7 @@ class Mollie extends PaymentModule
         }
 
         try {
-            $this->api = $apiKeyService->setApiKey($apiKey, $this->version, $subscriptionOrder);
+            $this->api = $apiKeyService->setApiKey($apiKey, $this->version, $subscriptionOrder, $environment);
         } catch (\Mollie\Api\Exceptions\IncompatiblePlatform $e) {
             $errorHandler = \Mollie\Handler\ErrorHandler\ErrorHandler::getInstance();
             $errorHandler->handle($e, $e->getCode(), false);
@@ -1301,6 +1301,7 @@ class Mollie extends PaymentModule
 
             $logger->error(sprintf('%s - General exception', self::FILE_NAME), [
                 'exceptions' => ExceptionUtility::getExceptions($e),
+                'mode' => $environment === Mollie\Config\Config::ENVIRONMENT_TEST ? 'test' : 'live',
             ]);
         }
     }