+++ src/Config/Config.php
@@ -75,6 +75,7 @@ class Config
         'multibanco' => ['pt'],
         'mbway' => ['pt'],
         'swish' => ['se'],
+        'bizum' => ['es', 'ad'],
     ];
 
     const SUPPORTED_PHP_VERSION = '5.6';
@@ -342,6 +343,7 @@ class Config
         'paybybank' => 'Pay by Bank',
         'multibanco' => 'Multibanco',
         'mbway' => 'MBway',
+        'bizum' => 'Bizum',
     ];
 
     public const LOG_SEVERITY_LEVEL_INFORMATIVE = 1;