+++ changelog.md
@@ -3,6 +3,7 @@
 # Changelog #
 
 ## Changes in release 6.4.0 ##
++ Improved API key validation
 + Added support for negative payment fees
 
 ## Changes in release 6.3.0 ##