+++ changelog.md
@@ -5,6 +5,7 @@
 ## Changes in release 6.4.0 ##
 + Improved API key validation
 + Added support for negative payment fees
++ Added Bizum payment method
 
 ## Changes in release 6.3.0 ##
 + Apple Pay Direct fixed